+++ lib/src/common/common.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 
 part 'smart_auto_suggest_controller.dart';
 part 'smart_auto_suggest_data_source.dart';
+part 'smart_auto_suggest_highlight_text.dart';
 part 'smart_auto_suggest_item.dart';
 
 const kPickerHeight = 32.0;