+++ lib/src/widgets/smart_auto_suggest_view.dart
@@ -802,26 +802,37 @@ class _SmartAutoSuggestViewListState<T>
                         itemCount: sortedItems.length,
                         itemBuilder: (context, index) {
                           final item = sortedItems.elementAt(index);
-                          return widget.itemBuilder?.call(context, item) ??
-                              SmartAutoSuggestBoxOverlayTile(
-                                subtitle: null,
-                                title: DefaultTextStyle.merge(
-                                  child: item.child ?? Text(item.label),
-                                  style: item.enabled
-                                      ? null
-                                      : TextStyle(color: disabledColor),
-                                ),
-                                semanticLabel: item.semanticLabel ?? item.label,
-                                selected: item.selected || widget.node.hasFocus,
-                                onSelected: item.enabled
-                                    ? () => widget.onSelected(item)
-                                    : null,
-                                tileColor: tileColor,
-                                selectedTileColor: selectedTileColor,
-                                selectedTileTextColor: selectedTileTextColor,
-                                tilePadding: tilePadding,
-                                tileSubtitleStyle: tileSubtitleStyle,
-                              );
+                          if (widget.itemBuilder != null) {
+                            return widget.itemBuilder!(context, item);
+                          }
+                          if (item.builder != null) {
+                            return Focus(
+                              child: item.builder!(context, searchText),
+                            );
+                          }
+                          return SmartAutoSuggestBoxOverlayTile(
+                            subtitle: null,
+                            title: DefaultTextStyle.merge(
+                              child: item.child ??
+                                  SmartAutoSuggestHighlightText(
+                                    text: item.label,
+                                    query: searchText,
+                                  ),
+                              style: item.enabled
+                                  ? null
+                                  : TextStyle(color: disabledColor),
+                            ),
+                            semanticLabel: item.semanticLabel ?? item.label,
+                            selected: item.selected || widget.node.hasFocus,
+                            onSelected: item.enabled
+                                ? () => widget.onSelected(item)
+                                : null,
+                            tileColor: tileColor,
+                            selectedTileColor: selectedTileColor,
+                            selectedTileTextColor: selectedTileTextColor,
+                            tilePadding: tilePadding,
+                            tileSubtitleStyle: tileSubtitleStyle,
+                          );
                         },
                       ),
                     );